api/image-classifier/requirements.txt
torch
transformers
Pillow
Flask
requests

